Remove axum from gencmd, fix ssh command
The SSH server IP in the generated ssh command is wrong, because we need
to get the server IP with `tcp_stream.local_addr()` instead of with the
listener. But with axum it's over complicated to do that, so we just use
hyper, as we already do in http.rs.